requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
imbalanced-learn
joblib
supabase
python-dotenv
groq
plotly
pdfplumber
reportlab
